backend/FAST-IBAN_Project/utils/requirements.txt
basemap==1.3.8
basemap_data==1.3.2
CairoSVG==2.7.1
Cartopy==0.22.0
imageio==2.34.1
matplotlib==3.7.2
xarray==2023.1.0
netCDF4==1.6.4
h5py==3.8.0
numpy==1.25.2
pandas==2.0.3
scipy==1.13.0
cdsapi>=0.7.4
pika>=1.3.0
python-dotenv>=1.0.0